pyproject.toml
[tool.poetry]
name = "estudando"
version = "0.1.0"
description = ""
authors = ["alexandrejastrow <alexandre.jastrow@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy = "^1.4.31"
email-validator = "^1.1.3"
requests = "^2.27.1"
orjson = "^3.6.7"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
